Update and rename tokenomics_farmers_grid_v1.md to tokenomics_farmers_grid_v1.1.md
updated # to be more precise why we need burning and minor language updates

## is unchanged and works well in my opinion

### 
- removed the details about selling order sizes, because it might limit us when the market demands flexibility from the market maker

The conversion of remaining TFTA could be instanced in 5% a month starting 01/2022

The min TFT price needs to be lower in my opinion. Would not tie it to another asset like gold because that adds unnecessary complexity. Gold may become more volatile in the next months because they are discovering certificates that are not covered by real reserves etc.
I would stick to the 0.15 we have now. If TFTA holders can sell for that, they are already making